+++ spec/spec_helper.rb
@@ -10,7 +10,7 @@ require 'rubygems'
 require 'bundler/setup'
 require 'vcr'
 require 'cgi'
-require 'ideal-mollie'
+require 'mollie'
 
 VCR.configure do |c|
   c.cassette_library_dir = 'spec/vcr_cassettes'